requirements.txt
Flask==2.0.1
requests==2.25.1
python-dotenv==0.21.1
gunicorn==20.1.0
